+++ example/lib/main.dart
@@ -1,18 +1,27 @@
+import 'dart:convert';
+
+import 'package:cryptography/cryptography.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:sprint_check/sprint_check.dart';
 import 'package:sprint_check/sprint_check_method_channel.dart';
 
-void main() {
-  runApp(const MyApp());
+Future<void> main() async {
+  WidgetsFlutterBinding.ensureInitialized();
+  // final cameras = await availableCameras();
+  // runApp(MyApp(cameras: cameras));
+  runApp(MyApp());
 }
 
 class MyApp extends StatelessWidget {
+  // final List<CameraDescription> cameras;
+  // const MyApp({super.key, required this.cameras});
   const MyApp({super.key});
 
   @override
   Widget build(BuildContext context) {
     return MaterialApp(home: Mypage());
+    // return MaterialApp(home: LivenessCheckScreen(cameras: cameras));
   }
 }
 
@@ -27,9 +36,10 @@ class _MypageState extends State<Mypage> {
   String _platformVersion = 'Unknown';
   final _sprintCheckPlugin = SprintCheck();
 
-  TextEditingController bvnController = TextEditingController(
+  TextEditingController identifierController = TextEditingController(
     text: "odejinmiabraham@gmail.com",
   );
+  TextEditingController bvnController = TextEditingController();
   @override
   void initState() {
     super.initState();
@@ -93,6 +103,36 @@ class _MypageState extends State<Mypage> {
                 ),
                 keyboardType: TextInputType.phone,
                 inputFormatters: [],
+                controller: identifierController,
+                onChanged: (value) {},
+                validator: (value) {
+                  if (value!.isEmpty) {
+                    return "Kindly input your identifier";
+                  }
+
+                  return null;
+                },
+              ),
+              SizedBox(height: 20),
+              TextFormField(
+                autofillHints: [AutofillHints.telephoneNumber],
+                decoration: InputDecoration(
+                  hintText: "22123456768",
+                  enabledBorder: const OutlineInputBorder(
+                    borderRadius: BorderRadius.all(Radius.circular(15.0)),
+                    borderSide: BorderSide(color: Color(0xFF6A6C6A)),
+                  ),
+                  errorBorder: const OutlineInputBorder(
+                    borderRadius: BorderRadius.all(Radius.circular(15.0)),
+                    borderSide: BorderSide(color: Colors.red),
+                  ),
+                  focusedBorder: const OutlineInputBorder(
+                    borderRadius: BorderRadius.all(Radius.circular(15.0)),
+                    borderSide: BorderSide(color: Color(0xFF6A6C6A)),
+                  ),
+                ),
+                keyboardType: TextInputType.phone,
+                inputFormatters: [],
                 controller: bvnController,
                 onChanged: (value) {},
                 validator: (value) {
@@ -109,7 +149,8 @@ class _MypageState extends State<Mypage> {
                   var response = await _sprintCheckPlugin.checkout(
                     context,
                     CheckoutMethod.bvn,
-                    bvnController.text,
+                    identifierController.text,
+                    bvn: bvnController.text,
                   );
                   showresult("response for the sdk: ${response}");
                   print("response for the sdk: ${response}");
@@ -122,7 +163,8 @@ class _MypageState extends State<Mypage> {
                   var response = await _sprintCheckPlugin.checkout(
                     context,
                     CheckoutMethod.nin,
-                    "odejinmiabraham@gmail.com",
+                    identifierController.text,
+                    nin: bvnController.text,
                   );
                   showresult("response for the sdk: ${response}");
                   print("response for the sdk: ${response}");
@@ -135,12 +177,27 @@ class _MypageState extends State<Mypage> {
                   var response = await _sprintCheckPlugin.checkout(
                     context,
                     CheckoutMethod.facial,
-                    "odejinmiabraham@gmail.com",
+                    identifierController.text,
                   );
                   showresult("response for the sdk: ${response}");
                   print("response for the sdk: ${response}");
                 },
                 child: Text("Start Face verification"),
+              ),
+              SizedBox(height: 20),
+              InkWell(
+                onTap: () async {
+                  var response = await _sprintCheckPlugin.checkout(
+                    context,
+                    CheckoutMethod.idcard,
+                    identifierController.text,
+                  );
+                  showresult("response for the sdk: ${response}");
+                  print("response for the sdk: ${response}");
+                  // startEncryption();
+                  // // decryptData();
+                },
+                child: Text("Start Id card verification"),
               ),
             ],
           ),
@@ -152,7 +209,7 @@ class _MypageState extends State<Mypage> {
   showresult(message) {
     showDialog(
       context: context,
-      barrierDismissible: false,
+      // barrierDismissible: false,
       builder: (BuildContext context) {
         return AlertDialog(
           content: Container(
@@ -174,5 +231,76 @@ class _MypageState extends State<Mypage> {
         );
       },
     );
+  }
+
+  // AES-GCM with 256-bit key
+  final algorithm = AesGcm.with256bits();
+  Future<void> startEncryption() async {
+    // AES-256 requires a 32-byte key
+    final keyBytes = utf8.encode('BaVkxaDFoNzI2U0FHa2o1OTJ2aytEeVY');
+    final secretKey = SecretKey(keyBytes);
+
+    // Generate a 12-byte random IV (nonce)
+    final nonce = algorithm.newNonce(); // 12-byte random nonce
+
+    var body = {"email": "odejinmiabraham@gmail.com", "password": "adeyemi"};
+    var jsonbody1 = jsonEncode(body);
+
+    // Calculate string length (number of characters)
+    final length = jsonbody1.length;
+
+    // Build PHP-style serialized string
+    final phpSerialized = 's:$length:"$jsonbody1";';
+
+    print(phpSerialized);
+    var unrfy1 = utf8.encode(phpSerialized);
+
+    // Encrypt
+    final secretBox = await algorithm.encrypt(
+      unrfy1,
+      secretKey: secretKey,
+      nonce: nonce,
+    );
+
+    // Build JSON structure similar to Laravel
+    final jsonResult = {
+      "iv": base64Encode(secretBox.nonce), // 12-byte IV
+      "value": base64Encode(secretBox.cipherText), // ciphertext
+      "mac": "", // Laravel leaves empty for GCM
+      "tag": base64Encode(secretBox.mac.bytes), // 16-byte tag
+    };
+    var jsonbody = jsonEncode(jsonResult);
+    print(jsonbody);
+    var unrfy = utf8.encode(jsonbody);
+    print(base64Encode(unrfy));
+
+    // --- Decrypt to verify ---
+    // final decrypted = await algorithm.decrypt(secretBox, secretKey: secretKey);
+    // print('Decrypted: ${utf8.decode(decrypted)}');
+  }
+
+  decryptData() async {
+    // AES-256 requires a 32-byte key
+    final keyBytes = utf8.encode('BaVkxaDFoNzI2U0FHa2o1OTJ2aytEeVY');
+    final secretKey = SecretKey(keyBytes);
+
+    var data =
+        "eyJpdiI6InVwZzNCUUxVMTJJd2l2emUiLCJ2YWx1ZSI6ImRGbXlUTTNyWXgwbVBuL1IwTFZUOTFkV2ZSUkRLWnNlVFJOaDNYdVBTdk9kYzY5L1hWWFp0d1djS09pTFJ3cXpwa01PeU5LOVhSQkZmTVFROG1vSnplTT0iLCJtYWMiOiIiLCJ0YWciOiI0WisvUlBySWRIMUU2R011S3ZnWWNRPT0ifQ==";
+    var unrfy = base64Decode(data);
+    print(unrfy);
+    var body = utf8.decode(unrfy);
+    print(body);
+    var bodyJson = jsonDecode(body);
+    print(bodyJson);
+    final decrypted = await algorithm.decrypt(
+      SecretBox(
+        base64Decode(bodyJson["value"]),
+        nonce: base64Decode(bodyJson["iv"]),
+        mac: Mac(base64Decode(bodyJson["tag"])),
+      ),
+      secretKey: secretKey,
+    );
+    print(utf8.decode(decrypted));
+    // print('Decrypted: ${utf8.decode(decrypted)}');
   }
 }