+++ example/windows/flutter/generated_plugin_registrant.cc
@@ -6,9 +6,15 @@
 
 #include "generated_plugin_registrant.h"
 
+#include <file_selector_windows/file_selector_windows.h>
+#include <permission_handler_windows/permission_handler_windows_plugin.h>
 #include <sprint_check/sprint_check_plugin_c_api.h>
 
 void RegisterPlugins(flutter::PluginRegistry* registry) {
+  FileSelectorWindowsRegisterWithRegistrar(
+      registry->GetRegistrarForPlugin("FileSelectorWindows"));
+  PermissionHandlerWindowsPluginRegisterWithRegistrar(
+      registry->GetRegistrarForPlugin("PermissionHandlerWindowsPlugin"));
   SprintCheckPluginCApiRegisterWithRegistrar(
       registry->GetRegistrarForPlugin("SprintCheckPluginCApi"));
 }