+++ CHANGELOG.md
@@ -8,3 +8,13 @@ example well details
 ## 0.0.3
 identifier added
 site url included
+
+##0.0.4
+bug fixed
+
+##0.0.5
+added facial
+
+##0.0.6
+direct checkout added
+you now pass your nin and bvn to the checkout method and it is optional